refactor(wa): separate client responsibilities and document ownership
