+++ internal/wa/client.go
@@ -2,7 +2,6 @@ package wa
 
 import (
 	"context"
-	"crypto/rand"
 	"fmt"
 	"os"
 	"strings"
@@ -12,15 +11,12 @@ import (
 	"github.com/mdp/qrterminal/v3"
 	"go.mau.fi/whatsmeow"
 	"go.mau.fi/whatsmeow/appstate"
-	waBinary "go.mau.fi/whatsmeow/binary"
 	waProto "go.mau.fi/whatsmeow/binary/proto"
 	"go.mau.fi/whatsmeow/proto/waE2E"
 	"go.mau.fi/whatsmeow/proto/waHistorySync"
-	"go.mau.fi/whatsmeow/proto/waSyncAction"
 	"go.mau.fi/whatsmeow/proto/waWeb"
 	"go.mau.fi/whatsmeow/types"
 	"go.mau.fi/whatsmeow/types/events"
-	"google.golang.org/protobuf/proto"
 )
 
 type Options struct {
@@ -281,224 +277,6 @@ func (c *Client) RemoveEventHandler(id uint32) {
 	cli.RemoveEventHandler(id)
 }
 
-func (c *Client) SendText(ctx context.Context, to types.JID, text string) (types.MessageID, error) {
-	c.mu.Lock()
-	cli := c.client
-	c.mu.Unlock()
-	if cli == nil || !cli.IsConnected() {
-		return "", fmt.Errorf("not connected")
-	}
-	msg := &waProto.Message{Conversation: &text}
-	resp, err := cli.SendMessage(ctx, to, msg)
-	if err != nil {
-		return "", err
-	}
-	return resp.ID, nil
-}
-
-func (c *Client) SendProtoMessage(ctx context.Context, to types.JID, msg *waProto.Message) (types.MessageID, error) {
-	return c.SendProtoMessageWithExtra(ctx, to, msg, "")
-}
-
-func (c *Client) SendProtoMessageWithExtra(ctx context.Context, to types.JID, msg *waProto.Message, mediaHandle string) (types.MessageID, error) {
-	c.mu.Lock()
-	cli := c.client
-	c.mu.Unlock()
-	if cli == nil || !cli.IsConnected() {
-		return "", fmt.Errorf("not connected")
-	}
-	if mediaHandle == "" {
-		resp, err := cli.SendMessage(ctx, to, msg)
-		if err != nil {
-			return "", err
-		}
-		return resp.ID, nil
-	}
-	resp, err := cli.SendMessage(ctx, to, msg, whatsmeow.SendRequestExtra{MediaHandle: mediaHandle})
-	if err != nil {
-		return "", err
-	}
-	return resp.ID, nil
-}
-
-// SendPoll builds a PollCreationMessage and sends it. selectable is the
-// maximum number of options a voter may pick (1 = single-select). The poll
-// can optionally be wrapped in an EphemeralMessage for disappearing chats.
-func (c *Client) SendPoll(ctx context.Context, to types.JID, name string, options []string, selectable int, ephemeral bool) (types.MessageID, error) {
-	c.mu.Lock()
-	cli := c.client
-	c.mu.Unlock()
-	if cli == nil || !cli.IsConnected() {
-		return "", fmt.Errorf("not connected")
-	}
-	var groupInfo *types.GroupInfo
-	if to.Server == types.GroupServer {
-		groupInfo, _ = cli.GetGroupInfo(ctx, to)
-	}
-	msg := buildPollCreationMessage(name, options, selectable, isCommunityAnnouncementGroup(groupInfo))
-	if ephemeral {
-		msg = wrapEphemeralPollMessage(msg)
-	}
-	resp, err := cli.SendMessage(ctx, to, msg)
-	if err != nil {
-		return "", err
-	}
-	return resp.ID, nil
-}
-
-func buildPollCreationMessage(name string, optionNames []string, selectableOptionCount int, toAnnouncementGroup bool) *waE2E.Message {
-	msgSecret := make([]byte, 32)
-	_, _ = rand.Read(msgSecret)
-	if selectableOptionCount < 0 || selectableOptionCount > len(optionNames) {
-		selectableOptionCount = 0
-	}
-	options := make([]*waE2E.PollCreationMessage_Option, len(optionNames))
-	for i, option := range optionNames {
-		options[i] = &waE2E.PollCreationMessage_Option{OptionName: proto.String(option)}
-	}
-	creation := &waE2E.PollCreationMessage{
-		Name:                   proto.String(name),
-		Options:                options,
-		SelectableOptionsCount: proto.Uint32(uint32(selectableOptionCount)),
-	}
-	msg := &waE2E.Message{
-		MessageContextInfo: &waE2E.MessageContextInfo{
-			MessageSecret: msgSecret,
-		},
-	}
-	switch {
-	case toAnnouncementGroup:
-		msg.PollCreationMessageV2 = creation
-	case selectableOptionCount == 1:
-		msg.PollCreationMessageV3 = creation
-	default:
-		msg.PollCreationMessage = creation
-	}
-	return msg
-}
-
-func isCommunityAnnouncementGroup(info *types.GroupInfo) bool {
-	return info != nil && info.IsAnnounce && info.IsParent
-}
-
-func pickOutboundPollCreation(msg *waE2E.Message) *waE2E.PollCreationMessage {
-	if msg == nil {
-		return nil
-	}
-	if msg.GetPollCreationMessage() != nil {
-		return msg.GetPollCreationMessage()
-	}
-	if msg.GetPollCreationMessageV2() != nil {
-		return msg.GetPollCreationMessageV2()
-	}
-	return msg.GetPollCreationMessageV3()
-}
-
-func wrapEphemeralPollMessage(msg *waE2E.Message) *waE2E.Message {
-	if msg == nil {
-		return nil
-	}
-	return &waE2E.Message{
-		EphemeralMessage:   &waE2E.FutureProofMessage{Message: msg},
-		MessageContextInfo: msg.MessageContextInfo,
-	}
-}
-
-// SendPollVote builds and sends a poll vote for the poll identified by
-// pollInfo (Chat, Sender, ID of the original PollCreationMessage). The
-// option names must match exactly the strings used in the poll.
-//
-// On migrated DM accounts, whatsmeow's SendMessage auto-rewrites the
-// destination from a phone-number JID to the corresponding LID. Pre-translate
-// DMs so the PollCreationMessageKey embedded by BuildPollVote matches the
-// chat/sender on the wire.
-func (c *Client) SendPollVote(ctx context.Context, pollInfo *types.MessageInfo, options []string) (types.MessageID, error) {
-	c.mu.Lock()
-	cli := c.client
-	c.mu.Unlock()
-	if cli == nil || !cli.IsConnected() {
-		return "", fmt.Errorf("not connected")
-	}
-	if pollInfo == nil {
-		return "", fmt.Errorf("poll info is required")
-	}
-
-	info := *pollInfo
-	info = rewritePollVoteInfoForLID(ctx, cli, info, c.resolvePNToLIDLocked)
-
-	msg, err := cli.BuildPollVote(ctx, &info, options)
-	if err != nil {
-		return "", fmt.Errorf("build poll vote: %w", err)
-	}
-	resp, err := cli.SendMessage(ctx, info.Chat, msg)
-	if err != nil {
-		return "", err
-	}
-	return resp.ID, nil
-}
-
-type pollVoteLIDResolver func(context.Context, *whatsmeow.Client, types.JID) types.JID
-
-func rewritePollVoteInfoForLID(ctx context.Context, cli *whatsmeow.Client, info types.MessageInfo, resolve pollVoteLIDResolver) types.MessageInfo {
-	if cli == nil || cli.Store == nil || cli.Store.LIDMigrationTimestamp <= 0 || resolve == nil {
-		return info
-	}
-	switch info.Chat.Server {
-	case types.DefaultUserServer:
-		info.Chat = resolve(ctx, cli, info.Chat)
-		if info.Sender.Server == types.DefaultUserServer {
-			info.Sender = resolve(ctx, cli, info.Sender)
-		}
-	case types.HiddenUserServer:
-		if info.Sender.Server == types.DefaultUserServer {
-			info.Sender = resolve(ctx, cli, info.Sender)
-		}
-	}
-	return info
-}
-
-// resolvePNToLIDLocked translates a phone-number JID to its LID counterpart
-// using the active session store; falls back to the input JID if no mapping
-// exists. Caller already holds (or doesn't need) c.mu.
-func (c *Client) resolvePNToLIDLocked(ctx context.Context, cli *whatsmeow.Client, jid types.JID) types.JID {
-	if cli == nil || cli.Store == nil {
-		return jid
-	}
-	pn := jid.ToNonAD()
-	if ownPN := cli.Store.GetJID().ToNonAD(); pn == ownPN {
-		if ownLID := cli.Store.GetLID().ToNonAD(); !ownLID.IsEmpty() {
-			return ownLID
-		}
-	}
-	if cli.Store.LIDs == nil {
-		return jid
-	}
-	lid, err := cli.Store.LIDs.GetLIDForPN(ctx, pn)
-	if err == nil && !lid.IsEmpty() {
-		return lid
-	}
-	info, err := cli.GetUserInfo(ctx, []types.JID{pn})
-	if err == nil {
-		if resolved := info[pn].LID.ToNonAD(); !resolved.IsEmpty() {
-			return resolved
-		}
-	}
-	return jid
-}
-
-// DecryptPollVote decrypts an incoming PollUpdateMessage event and returns
-// the SHA-256 hashes of the selected options. The caller is responsible for
-// matching those hashes back to option names.
-func (c *Client) DecryptPollVote(ctx context.Context, evt *events.Message) (*waE2E.PollVoteMessage, error) {
-	c.mu.Lock()
-	cli := c.client
-	c.mu.Unlock()
-	if cli == nil {
-		return nil, fmt.Errorf("whatsapp client is not initialized")
-	}
-	return cli.DecryptPollVote(ctx, evt)
-}
-
 func (c *Client) DecryptSecretEncryptedMessage(ctx context.Context, evt *events.Message) (*waE2E.Message, error) {
 	c.mu.Lock()
 	cli := c.client
@@ -520,96 +298,6 @@ func (c *Client) DeleteHistorySyncMedia(ctx context.Context, notif *waE2E.Histor
 		return nil
 	}
 	return cli.DeleteMedia(ctx, whatsmeow.MediaHistory, notif.GetDirectPath(), notif.GetFileEncSHA256(), notif.GetEncHandle())
-}
-
-func (c *Client) SendReaction(ctx context.Context, chat, sender types.JID, targetID types.MessageID, reaction string) (types.MessageID, error) {
-	c.mu.Lock()
-	cli := c.client
-	c.mu.Unlock()
-	if cli == nil || !cli.IsConnected() {
-		return "", fmt.Errorf("not connected")
-	}
-	resp, err := cli.SendMessage(ctx, chat, cli.BuildReaction(chat, sender, targetID, reaction))
-	if err != nil {
-		return "", err
-	}
-	return resp.ID, nil
-}
-
-func (c *Client) RevokeMessage(ctx context.Context, chat types.JID, targetID types.MessageID) (types.MessageID, error) {
-	c.mu.Lock()
-	cli := c.client
-	c.mu.Unlock()
-	if cli == nil || !cli.IsConnected() {
-		return "", fmt.Errorf("not connected")
-	}
-	resp, err := cli.SendMessage(ctx, chat, cli.BuildRevoke(chat, types.EmptyJID, targetID))
-	if err != nil {
-		return "", err
-	}
-	return resp.ID, nil
-}
-
-func (c *Client) DeleteMessageForMe(ctx context.Context, info types.MessageInfo, deleteMedia bool) error {
-	c.mu.Lock()
-	cli := c.client
-	c.mu.Unlock()
-	if cli == nil || !cli.IsConnected() {
-		return fmt.Errorf("not connected")
-	}
-	if info.Chat.IsEmpty() || strings.TrimSpace(string(info.ID)) == "" {
-		return fmt.Errorf("message chat and ID are required")
-	}
-	return cli.SendAppState(ctx, buildDeleteForMePatch(info, deleteMedia))
-}
-
-func buildDeleteForMePatch(info types.MessageInfo, deleteMedia bool) appstate.PatchInfo {
-	fromMe := "0"
-	if info.IsFromMe {
-		fromMe = "1"
-	}
-	sender := "0"
-	if !info.IsFromMe && !info.Sender.IsEmpty() && info.Chat.User != info.Sender.User {
-		sender = info.Sender.String()
-	}
-	return appstate.PatchInfo{
-		Type: appstate.WAPatchRegularHigh,
-		Mutations: []appstate.MutationInfo{{
-			Index:   []string{appstate.IndexDeleteMessageForMe, info.Chat.String(), string(info.ID), fromMe, sender},
-			Version: 2,
-			Value: &waSyncAction.SyncActionValue{
-				DeleteMessageForMeAction: &waSyncAction.DeleteMessageForMeAction{
-					DeleteMedia:      proto.Bool(deleteMedia),
-					MessageTimestamp: proto.Int64(info.Timestamp.UnixMilli()),
-				},
-			},
-		}},
-	}
-}
-
-func (c *Client) EditMessage(ctx context.Context, chat types.JID, targetID types.MessageID, text string) (types.MessageID, error) {
-	c.mu.Lock()
-	cli := c.client
-	c.mu.Unlock()
-	if cli == nil || !cli.IsConnected() {
-		return "", fmt.Errorf("not connected")
-	}
-	msg := &waE2E.Message{Conversation: proto.String(text)}
-	resp, err := cli.SendMessage(ctx, chat, cli.BuildEdit(chat, targetID, msg))
-	if err != nil {
-		return "", err
-	}
-	return resp.ID, nil
-}
-
-func (c *Client) Upload(ctx context.Context, data []byte, mediaType whatsmeow.MediaType) (whatsmeow.UploadResponse, error) {
-	c.mu.Lock()
-	cli := c.client
-	c.mu.Unlock()
-	if cli == nil || !cli.IsConnected() {
-		return whatsmeow.UploadResponse{}, fmt.Errorf("not connected")
-	}
-	return cli.Upload(ctx, data, mediaType)
 }
 
 func (c *Client) DecryptReaction(ctx context.Context, reaction *events.Message) (*waProto.ReactionMessage, error) {
@@ -733,127 +421,6 @@ func (c *Client) FetchAppStateEvents(ctx context.Context, name string, fullSync,
 	return cli.DangerousInternals().FetchAppState(ctx, appstate.WAPatchName(name), fullSync, onlyIfNotSynced)
 }
 
-func (c *Client) GetUserInfo(ctx context.Context, jids []types.JID) (map[types.JID]types.UserInfo, error) {
-	c.mu.Lock()
-	cli := c.client
-	c.mu.Unlock()
-	if cli == nil || !cli.IsConnected() {
-		return nil, fmt.Errorf("not connected")
-	}
-	return cli.GetUserInfo(ctx, jids)
-}
-
-func (c *Client) IsOnWhatsApp(ctx context.Context, phones []string) ([]types.IsOnWhatsAppResponse, error) {
-	c.mu.Lock()
-	cli := c.client
-	c.mu.Unlock()
-	if cli == nil || !cli.IsConnected() {
-		return nil, fmt.Errorf("not connected")
-	}
-	return cli.IsOnWhatsApp(ctx, phones)
-}
-
-func (c *Client) GetContact(ctx context.Context, jid types.JID) (types.ContactInfo, error) {
-	c.mu.Lock()
-	cli := c.client
-	c.mu.Unlock()
-	if cli == nil || cli.Store == nil || cli.Store.Contacts == nil {
-		return types.ContactInfo{}, fmt.Errorf("contacts store not available")
-	}
-	return cli.Store.Contacts.GetContact(ctx, jid)
-}
-
-func (c *Client) GetAllContacts(ctx context.Context) (map[types.JID]types.ContactInfo, error) {
-	c.mu.Lock()
-	cli := c.client
-	c.mu.Unlock()
-	if cli == nil || cli.Store == nil || cli.Store.Contacts == nil {
-		return nil, fmt.Errorf("contacts store not available")
-	}
-	return cli.Store.Contacts.GetAllContacts(ctx)
-}
-
-func (c *Client) ResolveLIDToPN(ctx context.Context, jid types.JID) types.JID {
-	if jid.Server != types.HiddenUserServer {
-		return jid
-	}
-	c.mu.Lock()
-	cli := c.client
-	c.mu.Unlock()
-	if cli == nil || cli.Store == nil || cli.Store.LIDs == nil {
-		return jid
-	}
-	pn, err := cli.Store.LIDs.GetPNForLID(ctx, jid.ToNonAD())
-	if err != nil || pn.IsEmpty() {
-		return jid
-	}
-	return pn
-}
-
-func (c *Client) ResolvePNToLID(ctx context.Context, jid types.JID) types.JID {
-	if jid.Server != types.DefaultUserServer {
-		return jid
-	}
-	c.mu.Lock()
-	cli := c.client
-	c.mu.Unlock()
-	return c.resolvePNToLIDLocked(ctx, cli, jid)
-}
-
-func BestContactName(info types.ContactInfo) string {
-	if !info.Found {
-		return ""
-	}
-	if s := strings.TrimSpace(info.FullName); s != "" {
-		return s
-	}
-	if s := strings.TrimSpace(info.FirstName); s != "" {
-		return s
-	}
-	if s := strings.TrimSpace(info.BusinessName); s != "" {
-		return s
-	}
-	if s := strings.TrimSpace(info.PushName); s != "" && s != "-" {
-		return s
-	}
-	if s := strings.TrimSpace(info.RedactedPhone); s != "" {
-		return s
-	}
-	return ""
-}
-
-func (c *Client) ResolveChatName(ctx context.Context, chat types.JID, pushName string) string {
-	fallback := chat.String()
-
-	if chat.Server == types.NewsletterServer {
-		meta, err := c.GetNewsletterInfo(ctx, chat)
-		if err == nil && meta != nil {
-			if name := NewsletterName(meta); name != "" {
-				return name
-			}
-		}
-	} else if chat.Server == types.GroupServer || chat.IsBroadcastList() {
-		info, err := c.GetGroupInfo(ctx, chat)
-		if err == nil && info != nil {
-			if name := strings.TrimSpace(info.GroupName.Name); name != "" {
-				return name
-			}
-		}
-	} else {
-		info, err := c.GetContact(ctx, chat.ToNonAD())
-		if err == nil {
-			if name := BestContactName(info); name != "" {
-				return name
-			}
-		}
-	}
-
-	if name := strings.TrimSpace(pushName); name != "" && name != "-" {
-		return name
-	}
-	return fallback
-}
-
 func (c *Client) GetGroupInfo(ctx context.Context, jid types.JID) (*types.GroupInfo, error) {
 	c.mu.Lock()
 	cli := c.client
@@ -904,98 +471,6 @@ func (c *Client) Logout(ctx context.Context) error {
 		return fmt.Errorf("not initialized")
 	}
 	return cli.Logout(ctx)
-}
-
-// SetProfilePicture sets the profile picture of the authenticated account.
-// avatar must be JPEG bytes; pass nil to remove the picture.
-// Returns the new picture ID assigned by WhatsApp.
-//
-// Uses DangerousInternals.SendIQ to send the w:profile:picture IQ stanza
-// without a "target" attribute, which is the correct format for updating
-// your own profile picture (as opposed to SetGroupPhoto which always sets target).
-func (c *Client) SetProfilePicture(ctx context.Context, avatar []byte) (string, error) {
-	c.mu.Lock()
-	cli := c.client
-	c.mu.Unlock()
-	if cli == nil || !cli.IsConnected() {
-		return "", fmt.Errorf("not connected")
-	}
-
-	var content any
-	if avatar != nil {
-		content = []waBinary.Node{{
-			Tag:     "picture",
-			Attrs:   waBinary.Attrs{"type": "image"},
-			Content: avatar,
-		}}
-	}
-
-	resp, err := cli.DangerousInternals().SendIQ(ctx, whatsmeow.DangerousInfoQuery{
-		Namespace: "w:profile:picture",
-		Type:      "set",
-		To:        types.ServerJID,
-		Content:   content,
-	})
-	if err != nil {
-		return "", err
-	}
-	if avatar == nil {
-		return "remove", nil
-	}
-	pictureID, ok := resp.GetChildByTag("picture").Attrs["id"].(string)
-	if !ok {
-		return "", fmt.Errorf("no picture ID in response")
-	}
-	return pictureID, nil
-}
-
-func (c *Client) GetProfilePictureInfo(ctx context.Context, jid types.JID, preview bool, existingID string) (*types.ProfilePictureInfo, error) {
-	c.mu.Lock()
-	cli := c.client
-	c.mu.Unlock()
-	if cli == nil || !cli.IsConnected() {
-		return nil, fmt.Errorf("not connected")
-	}
-	return cli.GetProfilePictureInfo(ctx, jid, &whatsmeow.GetProfilePictureParams{
-		Preview:    preview,
-		ExistingID: existingID,
-	})
-}
-
-func (c *Client) SetStatusMessage(ctx context.Context, msg string) error {
-	c.mu.Lock()
-	cli := c.client
-	c.mu.Unlock()
-	if cli == nil || !cli.IsConnected() {
-		return fmt.Errorf("not connected")
-	}
-	return cli.SetStatusMessage(ctx, types.SetStatusInput{Text: &msg})
-}
-
-func (c *Client) SetProfileName(ctx context.Context, name string) error {
-	c.mu.Lock()
-	cli := c.client
-	c.mu.Unlock()
-	if cli == nil || !cli.IsConnected() {
-		return fmt.Errorf("not connected")
-	}
-	if err := cli.SendAppState(ctx, appstate.BuildSettingPushName(name)); err != nil {
-		return err
-	}
-	if cli.Store != nil {
-		cli.Store.PushName = name
-	}
-	return nil
-}
-
-func (c *Client) GetBusinessProfile(ctx context.Context, jid types.JID) (*types.BusinessProfile, error) {
-	c.mu.Lock()
-	cli := c.client
-	c.mu.Unlock()
-	if cli == nil || !cli.IsConnected() {
-		return nil, fmt.Errorf("not connected")
-	}
-	return cli.GetBusinessProfile(ctx, jid)
 }
 
 // Reconnect loop helper.
